+++ examples/example_long_leeway_backtrack.py
@@ -40,6 +40,8 @@ text = [{'s': 'Observation', 'x': ilon, 'y': ilat, 'fontsize': 20, 'color': 'g',
 # Define domain of possible origin
 lons = np.arange(3.4, 5, .1/20)
 lats = np.arange(59.7, 60.8, .05/20)
+#lons = lons[0::2]  # using every second, due to memory limitation on CircleCI
+#lats = lats[0::2]
 corners = [lons[0], lons[-1], lats[0], lats[-1]]
 lons, lats = np.meshgrid(lons, lats)
 
@@ -53,7 +55,8 @@ density_backwards = od.get_histogram(pixelsize_m=5000).isel(time=-1).isel(origin
 density_backwards = density_backwards.where(density_backwards>0)
 density_backwards = density_backwards/density_backwards.sum()*100
 vmax = density_backwards.max()
-o.plot(background=density_backwards, clabel='Probability of origin [%]', text=text, corners=corners, fast=True, markersize=.5, lalpha=.02, vmin=0, vmax=vmax)
+o.plot(background=density_backwards, clabel='Probability of origin [%]', text=text, corners=corners,
+       fast=True, markersize=.5, lalpha=.02, vmin=0, vmax=vmax)
 os.remove(outfile)
 
 #%%
@@ -61,7 +64,7 @@ os.remove(outfile)
 o = Leeway(loglevel=50)
 o.add_reader([reader_norkyst, reader_arome])
 o.seed_elements(lon=lons, lat=lats, radius=0,
-                 time=start_time, object_type=object_type)
+                time=start_time, object_type=object_type)
 o.run(duration=duration, time_step=900, time_step_output=3600, outfile=outfile)
 print(o)
 
@@ -78,16 +81,20 @@ hit_start_lons = lon[hits, 0]
 hit_start_lats = lat[hits, 0]
 o_hit = opendrift.open(outfile, elements=hits)
 
-o.animation(compare=o_hit, legend=['Elements not hitting target', 'Elements hitting target'], fast=True, corners=corners, text=text)
+o.animation(compare=o_hit, legend=['Elements not hitting target', 'Elements hitting target'],
+            fast=True, corners=corners, text=text)
 
 #%%
 # .. image:: /gallery/animations/example_leeway_backtrack_0.gif
 
-o.plot(compare=o_hit, legend=['Elements not hitting target', 'Elements hitting target'], show_elements=False, fast=True, corners=corners, text=text)
+o.plot(compare=o_hit, legend=['Elements not hitting target', 'Elements hitting target'],
+       show_elements=False, fast=True, corners=corners, text=text)
 
 #%%
 # Plot the initial density of elements that actually hit the target after 24 hours. To be compared with the density figure from backwards simulation (see top)
 of = opendrift.open_xarray(outfile, elements=hits)
 density_forwards = of.get_histogram(pixelsize_m=5000).isel(time=0).isel(origin_marker=0)
 density_forwards = density_forwards.where(density_forwards>0)
-o_hit.plot(background=density_forwards/density_forwards.sum()*100, clabel='Probability of origin [%]', text=text, corners=corners, fast=True, markersize=.5, lalpha=.02, vmin=0, vmax=vmax)
+ratio = density_forwards/density_forwards.sum()*100
+o_hit.plot(background=ratio, clabel='Probability of origin [%]', text=text, corners=corners,
+           fast=True, markersize=.5, lalpha=.02, vmin=0, vmax=vmax)